+++ src/components/guess-the-guild/assign-logos/components/Draggable.tsx
