+++ src/components/guess-the-guild/assign-logos/components/GuildCardWithDropzone.tsx
