+++ src/components/guess-the-guild/assign-logos/components/SkeletonAssignLogos.tsx
