+++ src/components/guess-the-guild/assign-logos/components/SourceDropzone.tsx
