+++ src/components/guess-the-guild/assign-logos/components/TargetDropzone.tsx
