+++ src/components/guess-the-guild/components/ResultAlert.tsx
