+++ src/components/guess-the-guild/components/ScoreIndicator.tsx
